+++ src/libs/protocols/base.ts
@@ -1,5 +1,5 @@
 import { Transaction } from "@mysten/sui/transactions";
-import { TxSorSwap } from "../../types/aggregator";
+import { Config, TxSorSwap } from "../../types/aggregator";
 import { normalizeStructTag, parseStructTag } from "@mysten/sui/utils";
 import { SuiUtils } from "../../utils/sui";
 import { TransactionResultItem } from "../../types/sui";
@@ -8,21 +8,25 @@ export interface BaseContractParams {
   swapInfo: TxSorSwap;
   inputCoinObject: TransactionResultItem;
   currentAccount: string;
+  config: Config;
 }
 
 export abstract class BaseContract {
   protected swapInfo: TxSorSwap;
   protected inputCoinObject: TransactionResultItem;
   protected currentAccount: string;
+  protected config: Config;
 
   constructor({
     swapInfo,
     inputCoinObject,
     currentAccount,
+    config,
   }: BaseContractParams) {
     this.swapInfo = swapInfo;
     this.inputCoinObject = inputCoinObject;
     this.currentAccount = currentAccount;
+    this.config = config;
   }
 
   abstract swap(tx: Transaction): Promise<TransactionResultItem>;