+++ src/libs/swapWithRoute.ts
@@ -1,17 +1,19 @@
 import { Transaction } from "@mysten/sui/transactions";
 import { ProtocolContract } from "./protocols";
-import { TxSorSwap } from "../types/aggregator";
+import { Config, TxSorSwap } from "../types/aggregator";
 import { TransactionResultItem } from "../types/sui";
 
 export async function swapWithRoute({
   route,
   inputCoinObject,
   currentAccount,
+  config,
   tx,
 }: {
   route: TxSorSwap[];
   inputCoinObject: TransactionResultItem;
   currentAccount: string;
+  config: Config;
   tx: Transaction;
 }): Promise<TransactionResultItem | undefined> {
   let inputTokenObject = inputCoinObject;
@@ -23,6 +25,7 @@ export async function swapWithRoute({
       swapInfo: swap,
       inputCoinObject: inputTokenObject,
       currentAccount,
+      config,
     });
     const tokenOut = await contractInstance.swap(tx);
     inputTokenObject = tokenOut;